requirements.txt
Django==5.2.5
django-admin-sortable2==2.2.8
django-tinymce==4.1.0
environs==14.3.0
requests==2.32.4
pillow==11.3.0